Add graph rendering states
